openmp: Mark hwloc allocated data as such
Add new field in kmp_mem_desc_t to signal hwloc managed allocations.  It
seems there was no clever way of doing so but to add a field, as
otherwise it would have required creating a different set of allocators
which may require rewriting large parts of the already existing code.

Minor changes :
  - Keep the KMP log for the memory allocation